+++ components/x-audio/stories/expanded.js
@@ -1,0 +1,17 @@
+exports.title = 'App player (expanded)';
+
+exports.data = {
+	expanded: true,
+	playing: true,
+	loading: false,
+	title: 'Notre-Dame fire, Goldman slips, Netflix spend',
+	seriesName: 'FT News Briefing',
+	onPlayClick: () => {},
+	onPauseClick: () => {},
+	onCloseClick: () => {}
+};
+
+// This reference is only required for hot module loading in development
+// <https://webpack.js.org/concepts/hot-module-replacement/>
+exports.m = module;
+