+++ components/x-interaction/__tests__/x-interaction.test.jsx
@@ -99,6 +99,23 @@ describe('x-interaction', () => {
 			).toBe(10);
 		});
 
+		it('should update props of base using async updater function from action', async () => {
+			const Base = () => null;
+			const Wrapped = withActions({
+				foo: () => async ({bar}) => ({ bar: bar + 5 }),
+			})(Base);
+
+			const target = mount(<Wrapped bar={5} />);
+
+			await target.find(Base).prop('actions').foo();
+			target.update();
+
+			expect(
+				target.find(Base).prop('bar')
+			).toBe(10);
+		});
+
+
 		it('should wait for promises and apply resolved state updates', async () => {
 			const Base = () => null;
 			const Wrapped = withActions({
@@ -141,6 +158,44 @@ describe('x-interaction', () => {
 			).toBe(false);
 		});
 
+		it(`shouldn't set isLoading back to false until everything is finished`, async () => {
+			const Base = () => null;
+			const Wrapped = withActions({
+				foo: () => new Promise(resolve => {
+					setTimeout(resolve, 200, { bar: 10 });
+				}),
+			})(Base);
+
+			const target = mount(<Wrapped bar={5} />);
+			const promise1 = target.find(Base).prop('actions').foo();
+
+			await new Promise(resolve => {
+				setTimeout(resolve, 100);
+			});
+
+			const promise2 = target.find(Base).prop('actions').foo();
+			target.update();
+
+			expect(
+				target.find(Base).prop('isLoading')
+			).toBe(true);
+
+			await promise1;
+			target.update();
+
+			expect(
+				target.find(Base).prop('isLoading')
+			).toBe(true);
+
+			await promise2;
+			target.update();
+
+			expect(
+				target.find(Base).prop('isLoading')
+			).toBe(false);
+		});
+
+
 		it('should update when outside props change but prefer state changes', async () => {
 			const Base = () => null;
 			const Wrapped = withActions({
@@ -163,6 +218,98 @@ describe('x-interaction', () => {
 				target.find(Base).prop('bar')
 			).toBe(15);
 		});
+		
+		it('should pass changed outside props to state updaters', async () => {
+			const Base = () => null;
+			const Wrapped = withActions({
+				foo: () => ({ bar }) => ({ bar: bar + 5 }),
+			})(Base);
+
+			const target = mount(<Wrapped bar={5} />);
+
+			target.setProps({ bar: 10 });
+			target.update();
+
+			await target.find(Base).prop('actions').foo();
+			target.update();
+
+			expect(
+				target.find(Base).prop('bar')
+			).toBe(15);
+		});
+
+		describe('actionsRef', () => {
+			it('should pass actions to actionsRef on mount and null on unmount', async () => {
+				const actionsRef = jest.fn();
+
+				const Base = () => null;
+				const Wrapped = withActions({
+					foo() {},
+				})(Base);
+
+				const target = mount(<Wrapped actionsRef={actionsRef} />);
+
+				expect(actionsRef).toHaveBeenCalled();
+				expect(actionsRef.mock.calls[0][0]).toHaveProperty('foo');
+
+				target.unmount();
+
+				expect(actionsRef).toHaveBeenLastCalledWith(null);
+			});
+
+			it('should pass all actions for rewrapped components', async () => {
+				const actionsRef = jest.fn();
+
+				const Base = () => null;
+				const Wrapped = withActions({
+					bar() {},
+				})(withActions({
+					foo() {},
+				})(Base));
+
+				mount(<Wrapped actionsRef={actionsRef} />);
+
+				expect(actionsRef).toHaveBeenCalled();
+				expect(actionsRef.mock.calls[0][0]).toHaveProperty('foo');
+				expect(actionsRef.mock.calls[0][0]).toHaveProperty('bar');
+			});
+		});
+
+		it(`shouldn't reset props when others change`, async () => {
+			const Base = () => null;
+			const Wrapped = withActions({
+				foo: () => ({ bar: 10 }),
+				baz: () => ({ quux: 10 }),
+			})(Base);
+
+			const target = mount(<Wrapped bar={5} quux={5} />);
+
+			await target.find(Base).prop('actions').foo();
+			await target.find(Base).prop('actions').baz();
+			target.update(); // tell enzyme things have changed
+
+			expect(
+				target.find(Base).prop('bar')
+			).toBe(10);
+
+			expect(
+				target.find(Base).prop('quux')
+			).toBe(10);
+		});
+
+		it('should get default state from second argument', async () => {
+			const Base = () => null;
+			const Wrapped = withActions({}, {
+				foo: 5
+			})(Base);
+
+			const target = mount(<Wrapped />);
+
+			expect(
+				target.find(Base).prop('foo')
+			).toBe(5);
+		});
+
 	});
 
 	describe.skip('server rendering');