+++ components/x-interaction/src/Interaction.jsx
@@ -7,27 +7,54 @@ import { registerComponent } from './concerns/register-component';
 // use the class version for clientside and the static version for server
 const Interaction = typeof window !== 'undefined' ? InteractionClass : InteractionSSR;
 
-export const withActions = (getActions) => (Component) => {
+const invoke = (fnOrObj, ...args) => typeof fnOrObj === 'function'
+	? fnOrObj(...args)
+	: fnOrObj;
+
+export const withActions = (getActions, getDefaultState = {}) => (Component) => {
+	const _wraps = { getActions, getDefaultState, Component };
+
+	// if the component we're wrapping is already wrapped, we don't want
+	// to wrap it further. so, discard its wrapper and rewrap the original
+	// component with the new actions on top
+	if(Component._wraps) {
+		const wrappedGetActions = Component._wraps.getActions;
+		const wrappedGetDefaultState = Component._wraps.getDefaultState;
+
+		Component = Component._wraps.Component;
+
+		getActions = initialState => Object.assign(
+			invoke(wrappedGetActions, initialState),
+			invoke(_wraps.getActions, initialState)
+		);
+
+		getDefaultState = initialState => Object.assign(
+			invoke(wrappedGetDefaultState, initialState),
+			invoke(_wraps.getDefaultState, initialState)
+		);
+	}
+
 	function Enhanced({
 		id,
-		actions: extraActions,
+		actionsRef,
+		serialiser,
 		...initialState
 	}) {
-		// support passing actions to withActions as an object or a function
-		// that's called with the initial state
-		const actions = typeof getActions === 'function'
-			? getActions(initialState)
-			: getActions;
+		const actions = invoke(getActions, initialState);
+		const defaultState = invoke(getDefaultState, initialState);
 
 		return <Interaction {...{
 			id,
 			Component,
-			initialState,
-			// if extraActions is defined, those are from another level
-			// of wrapping with withActions, so those should take precedence
-			actions: Object.assign(actions, extraActions),
+			initialState: Object.assign({}, defaultState, initialState),
+			actionsRef,
+			serialiser,
+			actions,
 		}} />;
 	}
+
+	// store what we're wrapping for later wrappers to replace
+	Enhanced._wraps = _wraps;
 
 	// set the displayName of the Enhanced component for debugging
 	wrapComponentName(Component, Enhanced);
@@ -38,5 +65,6 @@ export const withActions = (getActions) => (Component) => {
 	return Enhanced;
 };
 
-export { default as hydrate } from './Hydrate';
-export { getInteractionSerialiser } from './concerns/serialiser';
+export { hydrate } from './Hydrate';
+export { HydrationData } from './HydrationData';
+export { Serialiser } from './concerns/serialiser';